openssl: disable -DOPENSSL_USE_APPLINK option for 32-bit
For 64-bit, it's disable by default. The reason for disabling it also for 
32-bit is that it breaks app linking against static OpenSSL libs, so that 
any OpenSSL function that require calls to C RTL functions would 
crash with NULL pointer dereference (f.e. when reading/writing 
`BIO_new_fd()` or `BIO_new_file()` objects), unless manually initialising 
the OPENSSL_UplinkTable[] pointer table beforehand (which doesn't 
seem like something documented or recommended by OpenSSL).
For dynamic OpenSSL libs, it could still be useful in theory, though 
in practice it is very fragile and cannot be enabled without also enabling 
it for static libs (unless doing two separate build passes for static and 
dynamic libs). So the current limitation is that dynamic libs are better 
be used with apps that are build with MinGW and therefore rely on 
the same C RTL. It'd be nice if a future OpenSSL release would offer 
a clean way to initialise C RTL function pointers, f.e. similar to 
`CRYPTO_set_mem_functions()`.