prepare for publishing binaries from non-Windows platforms [ci skip]
via Travis CI, either Linux or macOS. Created binaries are identical
to the Windows builds in configuration and functionality. The only
minor difference is the use of mingw-w64 GCC compiler instead of
clang for curl and libssh2. (build for these two components was
just switched to use clang at release 7.56.1 of curl).

For this to work reliably the macOS build had to be split into
separate jobs for 32 and 64-bit platforms to fit into the timeslot
(of 50 minutes).

Ref: https://travis-ci.org/vszakats/harbour-deps/builds/291490165