openssl: fix OpenSSL 1.1.1d build + updates for future versions
This patch drops the local override of a secure OpenSSL prefix
(`C:/Windows/System32/OpenSSL`) for the Windows platform and now
uses the fixed and now secure default of OpenSSL
(`C:/Program Files/OpenSSL' and
`C:/Program Files/Common Files/SSL`). This also made it possible
to drop the local patch that let ./Configure allow a custom
Windows prefix be passed.

Also adapted packaging logic to the new layout of `DESTDIR`, to
support spaces in filenames and added `ct_log_list.cnf` and
`*.dist` copies to the distribution package. This also allowed
to re-enable most disabled `shellcheck` warnings.

Finally add an OpenSSL 1.1.1d patch to fix `make install` with
prefixes with spaces in them (such as the new default prefix
when building for Windows) to work on *nix platforms with
cross-builds.

OpenSSL patch submitted upstream:
https://github.com/openssl/openssl/pull/9864