drop the effort of moving curl/libssh2 builds to cmake [ci skip]
This effort effectively stalled and it seems to be still quite far away
to gain the same control as Makefile.m32 (or let alone make it produce
a usable output). While being significantly more opaque, complex and slow.
For now, maintaining Makefile.m32 seems more effective than going with
cmake. This has to be revisited once Makefile.m32 goes away or if someone
with cmake expertise gives it another go. The advantage would be to use
a mainstream curl build-system.