pass secrets via heredocs instead of echo
This has the advantage of omitting secrets with -x shell option (command
tracing) enabled. Meaning:

- it is no longer necessary to ensure that -x is disabled for certain
  script sections (via `set +x`)
- consequently it is now unnecessary to create subshells to make sure
  a previous -x setting is restored after script sections
- we no longer need to rely on CI capabilities to hide secrets from the
  log output (though it's a good safety measure)
- secrets will not spill into logs even if running the script outside of
  a CI environment
- command tracing can now be enabled in sensitive scripts sections
  (i.e.: `_pkg.sh`, `_sign-code.sh`, `_sign-pkg.sh`, `_ul.sh`),
  making the build log output wholly transparent and auditable.

The downside is readability which might be improved by using
`exec <<EOF`; subject to future work. This method is required when there
are multiple secrets to be passed to a single command, but for now there
is no such case in curl-for-win builds.