_pkg.sh: check output instead of result code [ci skip]
A '|| true' had to be added to make 'set -e' mode happy and not fail
in case the virustotals servers are failing. That interfered with
the next 'if's where the result code was checked, which became
a constant 0 after that update. So let's just check if the captured
curl output is non-empty instead of looking at the result code.