add flavour/branch using Schannel as its sole TLS backend [ci skip]
Instead of using OpenSSL as a default backend, schannel builds will
tap on the OS's TLS/SSL capabilities, meaning the exact feature set will
depend on OS version (and possibly OS configuration settings).

This should generally not pose a problem when running on a fairly recent
(officially supported) Windows release (meaning Windows 10 or equivalent
server version).

The expected advantage of these builds is significantly smaller on-disk
and in-memory footprint, faster and simpler builds, and all the upsides
of avoiding a complex 3rd party library such an OpenSSL with its
vulnerabilities and attack surface. One specific vulnerability that this
will universally fix is CVE-2019-12572 which to this day is not fixed in
OpenSSL, and in fact OpenSSL denies this has anything to do with OpenSSL
itself, so there is little chance this will ever get addressed.

This branch also allows to go back to the curl-for-win promise of not using
any local patches, by dropping a necessary OpenSSL patch to be able to
mitigate said CVE somewhat better. Another decisive moment was OpenSSL's
recently laid out plans that made it unlikely that this, or other pending
TLS/SSL/QUIC-related bits will get the expected attention in the
foreseeable future.

Ref: https://blog.mirch.io/2019/06/10/cve-2019-12572-pia-windows-privilege-escalation-malicious-openssl-engine/
Ref: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2019-12572
Ref: https://github.com/openssl/openssl/issues/17151
Ref: https://github.com/openssl/openssl/pull/16445
Ref: https://github.com/openssl/openssl/issues/9520