openssl.sh: force compatiblity with old/unsupported x86 Windows versions
OpenSSL started to link against Vista-specific symbols, thus breaking
compatibility with Windows XP when built using modern toolchain defaults.

This patch forces building OpenSSL for pre-Vista (XP) platforms on the
x86 (Intel 32-bit) architecture. Assuming that up-to-date environments
would normally use x64, x86 seems okay to be tweaked for old/legacy OS
versions. But, generally this isn't a good idea and also difficult to
test for regressions or other gotchas. So consider this experimental,
if there are reports of issues, this patch may be reverted.

Reported-by: Fierelier on github
Fixes #32