revert to openssl 3.0.2
OpenSSL 3.0.3 silently introduced an internal change that replaced
str[n]icmp() with str[n]icmp_l() throughout its complete codebase on the
Windows platform.

The new function is locale-dependent, which resulted in various fallouts
from 'i' not matching 'I' with Turkish locale to string matching failing
completely on some systems (notably Windows 7), thus not loading the
random generator engine and possibly other issues. This includes cases
using pure 7-bit ASCII (US-ASCII, English alphabet) as inputs.

This patch is a temporary solution and is planned to be addressed
long-term in a future commit.

It is recommended to upgrade to this curl-for-win build, but it should
be noted that this also reintroduces two OpenSSL vulnerabilities that
were fixed with 3.0.3, one with Moderate and one with Low severity [1].

[1] https://www.openssl.org/news/secadv/20220503.txt

Ref: https://github.com/openssl/openssl/issues/18322

Fixes https://github.com/curl/curl/issues/8849