stop distributing binaries not necessary for curl-for-win [ci skip]
- brotli: brotli.exe, libbrotli*.dll, implibs
- libidn2: idn2.exe
- openssl: openssl.exe, libcrypto-3[-x64.dll, libssl-3[-x64].dll, implibs,
           *.cnf
- in case of OpenSSL, a second, minor './Configure' patch had to be added
  to allow the 'no-apps' option to disable building 'openssl.exe'. Building
  'openssl.exe' and the DLLs have been broken on the macOS platform for
  a while, and made it impossible to perform tests locally.
- delete DLL configuration options where the DLL is not distributed.

curl did not need any of these to run, it is a self-contained binary.
libcurl[-x64].dll is a self-contained binary as well, that includes all
dependencies, so there is no need for further DLLs. We continue to
distribute all static libraries used to build these binaries. These can
be used to link libcurl statically to any user program (within limitation
of their licenses.)