libssh2: switch to autotools build [ci skip]
It did not become simpler, but at least now we are using a maintream
build. Hard to tell if CMake would any better.

I tried to make sure not to lose anything compared to the Makefile.m32
builds. One difference is that the new one will use 'long long' and
strtoll() for 64-bit numbers (such as the size of transferred files),
instead of '__int64' and strtoi64() as previously. This hopefully will
not make this old issue re-appear, but please report it if it does: [1]

[1] https://github.com/curl/curl/issues/858

While testing it turned out that libssh2 does not support
ed25519/curve25519 features with LibreSSL. LibreSSL has the
X25519_keypair() and X25519() functions to deal with curve25519.
Definitely something to note and a big downside when using LibreSSL
at the moment.

libssh2 autotools doesn't correctly recognize such basic component as
zlib just by specifying a proper path with its documented --with-libz-prefix
option. It also needs to be manually fed with the include and lib dirs.
A little less unexpected is that it also doesn't recognize OpenSSL in
x64 mode, since OpenSSL decided to rename the lib directory to lib64.
Apparently autotools is not aware of that, or I'm holding it wrong.