libressl.sh: disable __MINGW_USE_VC2005_COMPAT option [ci skip]
64-bit time_t would be cool for x86 builds, but it needs to be added
universally to all components linked together, and this appears to be
a long shot. Mixing objects with different settings may break the ABI
between them, so let's skip this for now. The recommended path is to
migrate to x64, which has this by default.

Background info:
https://github.com/libressl-portable/portable/blob/b2b47a7d113c2ac0e574ea4c3528780c30ef1aab/README.windows#L26-L43

MSYS2 has this enabled for OpenSSL:
https://github.com/msys2/MINGW-packages/blob/7184858fd44e5c0c8f31024064157c41c5374c68/mingw-w64-openssl/PKGBUILD#L112