fixups for openssl/libressl destpath normalization [ci skip]
I may abandon the idea to retain the OpenSSL 3.x feature of storing x64
static libs inside lib64 (and not lib), as it was ever before and as it
is with most packages built for Windows. It breaks all auto-detection
and needs constant fiddling to make it work. It's also difficult to say
what advantage this brings and how it fits into the bigger picture, e.g.
ARM64 builds. Will that also be put in lib64, or will it have a different
subdir? For the goal of uniformizing Windows with Unix (?), this seems
to have too much downsides.