add unified, single-file distro package [ci skip]
Single-file per target (win64, win32) curl-for-win binaries. The files
contain curl.exe, curl DLL + implib, docs and C headers + static libs
for curl and all of its dependencies, along with LICENSE/NEWS text files
for each dependency (also some extra docs for libssh2.)

These two files are included in the all-mingw*.zip package, and also
downloadable as an artifact from the CI page. The idea is that this
replaces the per-dependency packages in which curl components were
historically distributed in curl-for-win. This makes everything simpler,
especially for those who are looking for the static libs. Now they are
in one downloadable file instead of a bunch (8-10) of them. The new
files are named 'curl-uni-*', mainly to not collide with the single-file
curl package. Old-style packages will be discontinued, at which point
'curl-uni-*' will likely be renamed to simply 'curl-*'.

There is also some space/download-size savings with the unified packages.

This is in addition to size savings introduced recently by dropping
unused DLLs and EXEs, and enabling -O3 optimization for all dependencies.
This counteracts the slight size increase caused by HTTP/3 (aka QUIC)
support, introduced earlier today.

This will also need some updates on the server-side for publishing and
and a small update for the https://curl.se/windows/ page and scripts, by
deleting the individual dependency download links.