curl: abandoning efforts to move to autotools or CMake [ci skip]
After spending possibly weeks and multiple false starts throughout several
years attempting to move from Makefile.m32 to either autotools or CMake,
I'm admitting defeat and giving up.

It will be simpler to maintain Makefile.m32 for the next 20 years, than
to pull of this one-time migration. I cannot figure out how to make these
work for this specific cross-build scenario. I'm not sure it's even possible,
though it's also possible that I'm missing some fundamental configuration
with either build systems. With the current base settings they are both fatally
broken with linking a curl/libcurl binary at all. Not to say the falsely
detected components, the non-detected features and components, and the ones
that are detected, but misconfigured in various ways. Automation and hiding
things is good when it works, but when it doesn't, it's giving an extreme
hard time, while also not offering any (or any reasonable) manual override.
Documentation is almost non-existent, behaviour and naming things is widely
inconsistent and the source code of both of these tools is borderline
unreadable (often even unfindable.) It's mostly unclear which component is
responsible for what behaviour and at which stage of a multi-layered,
rich-with-code-generation, build process. Trial and error is cumbersome,
because running and re-running the slow autodetection process plus the whole
build has be go through every time, often for multiple projects when testing
various combinations.

Most of the time even essential info is missing such as if an input value is
expected to be a directory or a filename, an absolute or relative one, with
or without a prefix or extension, or ending-slash, or is it really a list,
separated by some character. If it's a directory, is pointing to the lib dir,
the project root dir, the header dir or some other variation. There is no way
to know, and there are no meaningful error messages either.

I'm sure autotools works for Unix and CMake mostly works for Windows, but
none seem to be able to correctly do a cross-build, especially not with clang.
For example autotools uses the regexp "cl*" to detect MSVC, which accidentally
also matches "clang", even though "clang" behaves like mingw/gcc. clang has
been around for 5 years (for Windows), at least.

Maybe one day there will be a build-system that is declarative and can also
execute the actual build without outsourcing it to the next tool. And one that
maybe doesn't test every Unix header that ever existed in the last 50 years,
when the build target is Windows. Then exits with a "fatal warning" about not
finding the system library 'ws2_32', or libz.

Another option is to give up on clang, switching back to gcc, which would fix
some of the issues, which in turn may make them produce an output.