curl.sh: use Makefile.m32 files directly [ci skip]
Without jumping through `Makefile` in the root. It makes things even
simpler and more straighforward. The `Makefile.m32` logic may even be
deleted from Makefile.dist to avoid confusion.

For a full tool_hugehelp.c, a release source tarball is necessary,
but the build also works with raw source snapshots, in that case with
a hugehelp placeholder.