curl: add experimental CMake build [ci skip]
Finally managed to find the key reason of the CMake build failures
with unresolved symbols (with both ld and llvm's lld), which were
obviously present in the list of libraries. After weeding out some
libraries that were picked up from system locations of the native
OS (e.g. /usr/local/), it turned out that the problem is that lib
files are not listed in the order their symbols are getting resolved
and/or there could possibly be some cases of circular dependency
(which is perfectly normal). In such case the solution is normally
to repeat the affected libraries twice at different positions, or,
a less effective option is to use the --start-group/--end-group
options to group them.

The order of the libs probably depend on the CMake build files
or CMake's internal detection logic. Maybe there is a way to control
it, or to insert libs at the end of the command-line, or at specific
places, but I failed to figure out how. Libs added as part of the
`CMAKE_*_LINKER_FLAGS[_INIT]` option, all seem to be added to the
beginning of the command-line, like any other custom option.

The "ugly hack" that resolved this problem was adding
`-Wl,--start-group` in the `CMAKE_*_LINKER_FLAGS` option so the
linker will make sure to resolve missing symbols in multiple passes
in all libs present on the command-line, regardless of their order
and interdependencies. A cosmetic fallout is that the missing
`--end-group` causes a linker warning. But, it makes the linker
resolve all symbols.

So finally a curl .exe and .dll is produced. The result is not yet
on par with production. NGTCP2, QUIC in OpenSSL, LIBIDN2 aren't
detected and even result in errors. Some internal features, like
case-insensitive compare are also mis-detected. libcurl.dll has
all (including dependencies) symbols exported because the .def
file provided is ignored, and curl.exe is not standalone because
BUILD_SHARED_LIBS also implied that curl.exe should be using it.

Any suggestion is welcome that may help getting closer on fixing
these.

[ Possibly even better would be the fix the autotools cross-build
issues, as this build system would open up the way to build curl
with rustls and other interesting options. ]