curl-cmake.sh: fix 3 critical binary creation issues [ci skip]
- fix ignored .def file when creating libcurl.dll. Now it is honored,
  so, only curl public symbols are exported from the DLL.
- fix missing libcurl static library. The static libcurl library is
  now built.
- fix curl.exe being built against libcurl.dll. After this patch curl.exe
  is linked against the static libcurl library, making it a standalone
  .exe.

The price for these fixes is slower builds. CMake has to be initialized
twice with one build pass after each to achieve the above results.
It's not tragic, but certainly not optimal. As demonstrated by Makefile.m32,
it's technically possible to build shared and static libs in a single pass,
and some curl dependencies do support that with their CMake setup, so maybe
one day curl's CMake will also receive a patch implementing it.