curl: fix to strip non-deterministic lld metadata [ci skip]
When using UCRT with clang, we must use LLVM's lld linker. But,
the binutil strip tool does not strip the `.buildid` section generated
by lld, which happens to contain a timestamp. LLVM (as of 13.0.1) doesn't
seem to offer an external tool to strip these [1] (or I did not find it),
so let's use the lld `-s` option for that, and skip calling binutil strip.

[1] It has an llvm-objcopy, which has some degree of Windows/COFF support.
Yet, I could not make it work, with this error:
  error: option not supported by llvm-objcopy for COFF