delete pkgconfig files [ci skip]
Beside some minimal useful information, the .pc files under the pkgconfig
directories generally do not help linking libcurl to an application.
The libpaths are absolute and *nix-specific, and unlikely to be correct in
any actual Windows development environment. Sometimes the included libname
is out of sync with the actual libnames on disk. When linking curl EXE and
libcurl DLL as part of the curl-for-win build process via the autotools
(./configure) method, these .pc files are picked up by the build process
and breaking it in various unexpected ways. Some components (i.e. nghttp3
and ngtcp2) explicitly require a pkgconfig subdir to exist, but they also
fail when trying to use them. To avoidsuch issues and the resulting confusion,
I opted to delete all .pc files and to configure dependencies manually
instead. The CMake build process is not affected, and Makefile.m32 never
relied on pkgconfig, so it is not affected either.