curl: enable HAVE_SIGNAL (m32, cmake) [ci skip]
Due to different detection methods, HAVE_SIGNAL is detected by autotools,
but not by CMake. signal() isn't a thing on Windows, so the correct
config would be !HAVE_SIGNAL, but instead of going through the pain of
fixing this in autotools (which technically is correct, because a
signal() function is indeed available in mingw-w64), let's enable it in
cmake and m32.

HAVE_SIGNAL enables an extra field (`prev_signal`) in the `UrlState`
struct. This field is not used even in autotools builds, it only consumes
some extra bytes. The significance of this is that with this field
enabled in all three build systems, it helps compiling bit-by-bit
identical .o output.

This might be fixed upstream by syncing the #if conditions in `urldata.h`
and `transfer.c`.