libssh2.sh: manually enable -DLIBSSH2_CLEAR_MEMORY [ci skip]
This option serves the same purpose as `-DHAVE_DECL_SECUREZEROMEMORY=1`,
but for the WinCNG (Schannel) and mbedTLS backends
(as of libssh2 1.10.0). It's also set wrongly by autotools, alongside
this warning:
```
configure: WARNING: secure clearing/zeroing of memory is not supported by the selected crypto backend
```

The macro guards SecureZeroMemory() calls, which is in fact available.

So let's enable it manually for all configurations.

OpenSSL/LibreSSL relies on HAVE_DECL_SECUREZEROMEMORY. It's an upstream
fix opportunity to delete this setting and replace these direct Win32
API calls with _libssh2_explicit_zero(), which handles this task based
on the HAVE_DECL_SECUREZEROMEMORY and other macros.