autotools/openssl: move build files under a subdir [ci skip]
Following similar cmake change c4d0df149987a4fc936ea7925b76f2b4e82d36f0,
with similar benefits.

For libressl.sh, this changes the generated binaries because __FILE__
values are different in an out-of-tree build.

Also do this for openssl, which supports out of tree builds even though
it is not autotools-based.

After this patch only curl-m32.sh does not support out of tree builds
and a DESTDIR, and therefore multiple targets in parallel.