fix dll/exe timestamps in non-zero timezone [ci skip]
On machines with a non-zero timezone _peclean.py miscalculated the
timestamp written into EXE/DLL files and breaking reproducibility.
This has been fixed by omitting extra timestamp manipulation and using
the unix timestamp returned by getmtime() as-is.

To avoid this issue in the future, log the reference file timestamp
in UTC and also show the timestamp that ended up in the EXE/DLL (also
in UTC). _peclean.py is now also showing the date in UTC to match them.

The CI machine has UTC timezone set, so it wasn't affected.