add BoringSSL support [ci skip]
Its downside is that this isn't supposed to be used as a dependency, has
no versioning or formal releases. TLS-SRP also isn't supported, but its
advantages outweigh it.

A smaller downside is that it uses pthread even on Windows, which has an
issue with static linking, possibly in combination with UCRT builds with
the cross-toolchain I tested it with. With this fixed, it's the most
promising alternative TLS backend. Binary size is also 2MB lower for both
curl EXE and DLL.

curl needs two patches, one merged [1], one fixing new iterations of a
known issue [2] pending.

[1] https://github.com/curl/curl/pull/9109
[2] https://github.com/curl/curl/pull/9110

curl would probably use a third patch from a CMake expert: ngtcp2
detection is broken, so we need to use an local workaround.

curl autotools doesn't support BoringSSL because I've found no way to
tell libtool to link pthread DLL instead of the static version.