curl-autotools.sh: use trick to link static/shared pthread [ci skip]
Regular methods don't work with libtool to control static/shared linkage
of libwinpthread, so resort to a trick and pass the .a file as an input
file with absolute path. This works, however we must keep track of where
this lib is on disk within the toolchain tree structure. Yet another
trick that makes autotools builds fragile and complex.

This trick is required regardless of toolchain or compiler with
autotools (i.e. not only necessary for the Homebrew mingw-w64
workaround).