fix C++ configuration and thus BoringSSL on CI [ci skip]
Turns out I overlooked it and std::thread support is in fact present in
Debian mingw-w64. Fixing the configuration to use it fixed BoringSSL
builds.

It means BoringSSL builds are ready for production and the only pending
issue is on macOS, where the pthread static linking issues occurs. It
can be fixed by using llvm-mingw on that platform.