stop using --preserve-dates strip option [ci skip]
This is unnecessary as we are explicitly setting the timestamp of these
binaries anyway in a subsequent step.

It also causes a problem with llvm's strip/objcopy tool, which doesn't
support it (as of today), leading to this fatal error:

  `aarch64-w64-mingw32-strip: error: option not supported by llvm-objcopy for COFF`

Removing the option makes it work as expected.

Also delete the workaround that disabled `strip` with llvm-mingw as a
workaround for the error. This also fixes inflated BoringSSL .a files.

TODO: Re-evaluate this fix: 27546d78c49704d332f7728e021c16d11c94849d