_build.sh: decouple lld usage from ucrt [ci skip]
It means that lld (and llvm-strip/llvm-objdump) are now always used with
clang, even when creating mscvrt builds. Before this patch, binutils was
used in that case, but these tools turned out to be working fine (as
tested with llvm/clang 13/14), so I feel comfortable to make this switch
now. This makes configuration cleaner. ld is now used with gcc only.