_build.sh: move -fno-ident C flag to gcc only [ci skip]
As of clang 13/14 this option seems to be a no-op for Windows COFF
output. clang doesn't write compiler info into these objects with or
without it. Documentation doesn't speak about any platform-specific
differences here, so leave a command about it, in case it needs to be
re-enabled in the future.

This also silences "unused option" warnings where this option ended up on
a linker command-line.

gcc 12.1.0 still needs it.