_build.sh: create and use llvm-windres alias on Linux [ci skip]
On Linux, the llvm-windres alias is missing, and in that case we have
been reverting to use binutils windres. This patch changes this to
create a local llvm-windres alias (to llvm-rc) and use that instead.

This means that we're back to using llvm-windres with clang on all
platforms.

With one exception: zlib, which will be tackled in a next commit.