add source tarball URLs to manifests [ci skip]
Append tham after the version number in per-package manifests
`BUILD-MANIFEST.txt`, add a new file `urls.txt` to the upload package
with '<package-name> <version> <url>' line format.

The latter is an extension of existing `build.txt`, and once we teach
curl-www how to parse it, we can drop the old format (and maybe re-use
the filename.)

The goal is making it easier to identfy/find the source components used
for building curl-for-win.