curl: sync OS string for all three build methods [ci skip]
Apply upstream patch [1] [2], which unconditionally sets the OS string to
the cross-build target value with CMake. For Makefile.m32, use the new
option to override the OS string to the triplet with an extra build
option.

The internal switch `CW_DEV_CROSSMAKE_REPR` no longer has an effect on
this. And, curl-for-win could get rid of two local (disabled-by-default)
patches.

[1] https://github.com/curl/curl/commit/ca73991774b93471fa09fc84790e101378786470
[2] https://github.com/curl/curl/pull/9117