boringssl.sh: fix cross-build-host reproducibility [ci skip]
llvm-strip has a list of issues that leave build-host specific bytes and
data in libcrypto.a and libssl.a. Resolve this by additionally running
them through binutils strip.

In that case also include the binutils version in the manifest.

Hopefully this gets fixed by future versions of llvm-strip. It's also a
prerequisite for enabling ASM for ARM64 boringssl builds.

Also change binutils version detection to use 'strip' instead of 'ar'.
The detected version is the same, and for non-gcc (clang) builds, the
only gcc tool we point to is 'strip'.

Also delete 'versuffix' from binutils. It's not an llvm-mingw-build-
specific version number and always depend on the system mingw-w64
installation. It's boringssl-specific.