nghttp2.sh: update accidental change to its latest state [ci skip]
It was an experiment to build nghttp2 apps, but after resolving a list
of issues, it turned out they most likely require more dependencies to
actually build. I'm committing this for future reference, before
deleting it in a next commit.

Upstream PR opened for an issue found in the process:
https://github.com/nghttp2/nghttp2/pull/1755