curl: synchronize libcurl.a output across build systems [ci skip]
By using a custom-made AR tool wrapper and a local script, normalize
libcurl.a output by sorting its objects, and making their names the
same across the three build systems, and thus making all three build
systems create a bit-by-bit identical libcurl.a.

This feature can be enabled via CW_DEV_CROSSMAKE_REPRO=1 as an
experiment to make build differences easier to sport via diff tools.

Using this method means that the final EXE/DLL are built from the
normalized libcurl.a files, so it's not an post-build transformation.
This also helps bringing EXE/DLL output closer to each other (we will
need to synchronize lib order in a future step.)

In general this method may be useful in other cases, where our only
current option is patching the build system. It's already used with
zlib's CMake to make it work with llvm-rc.

Another option is to migrate all libcurl.a post-processing into the
wrapper (keeping in mind that some build tools will apply runlib and
strip on the AR-wrapper output.)