cmake: delete underscore prefix from CFLAGS/LDFLAGS var names [ci skip]
The goal is to sync variables names used with cmake, autotools and other
builds system's initialization. Notice that with cmake, we never export
CFLAGS and LDFLAGS variables. We pass them via CMAKE_ variables.

(The reason for that is that I could not make CMake recognize an exported
CFLAGS variable, though I might be holding it wrong.)