_build.sh: always fill _BUILD_HOST [ci skip]
It wasn't filled when running on Windows and unrecognized platforms.
It's best to always fill it though, because even on Windows, a cross
build is a common scenario (e.g. building x64 on x86, or vice versa,
or building ARM64 on any an Intel machine, etc.)

This fixes builds running on Windows and may fix builds running on
not explicitly supported (unrecognized) build hosts.

On unrecognized platforms, we make a lazy effort to detect the build
host triplet. This is most likely wrong. It will need actual tests
to see whether omitting this `./configure` option would work better.