drop Windows XP compatibility [ci skip]
This affects x86 builds, the rest already required Vista.

MS released Vista 15 years ago in 2006. They EOLed XP in 2014 and even
its embedded iterations in 2019.

curl/libcurl continues to support XP, so if you need compatibility, it's
possible to build it, also with the previous commit of these scripts or
by re-applying this patch. You can keep using existing downloads on XP
of course.

After this patch libidn2 becomes optional. We preferred it so far to
avoid depending on Normaliz.dll, which was not available on XP and also
disappeared from MS servers as a installable option.