add SPDX-License-Identifier tags [ci skip]
Except for:
- .pem and .asc files. Found no guidance on what to do with those.
- .patch files. Seem labour intensive to maintain a license header for
  these (and with what content), in case this is necessary. Maybe
  better to not use any to avoid the hassle.

This repo managed to stay simple and not use subdirectories (especially
not fixed-name ones), so adding a LICENSE subdir with local copies of
used licenses is not on the table for REUSE-compliance. Same goes for the
.reuse subdirectory. It'd be great have a single-file-in-the-root method
to fix these up in a less obtrusive way.

If this catches on, it might make sense to replace freeform Copyright
headers with `SPDX-FileCopyrightText: [<year>] <name>` tags in a future
commit.