move back to WinIDN (from libidn2) by default [ci skip]
Switch (back) to using WinIDN instead of using libidn2. This is now
possible now that we require Vista, where the WinIDN subsystem is
always available out of the box.

This change:
- Ensures to use IDNA translations uniform with other Windows apps.
- Drops the libidn2 static lib dependency.
- Makes binaries smaller.
- Makes builds quicker.
- Switches from the IDNA 2008 standard to IDNA 2003 Transitional.

Alternate "big" builds continue to use libidn2.

More information on WinIDN/libidn2 differences:
- https://curl.se/docs/url-syntax.html#idna
- https://pypi.org/project/idna/

In isolated CLI tests I could not make IDN translations work, so it is
possible that this change is entirely academic, or that I'm missing some
critical details here:

> curl-winidn.exe -v https://é.com/
curl: (3) Failed to convert é.com to ACE; No Unicode translation.
> curl-libidn2.exe -v https://é.com/
curl: (3) Failed to convert é.com to ACE; could not convert string to UTF-8