curl-autotools.sh: use non-patch trick to skip building a shared curl.exe [ci skip]
Instead of patching the generated root Makefile, call lib/Makefile
directly to build libcurl.dll, without the DLL-based curl.exe (which
fails badly). This skips root Makefile build steps (e.g. installing
headers and a `share` directory), but we do those anyway in the next
step when building everything in static mode, so it not an issue and
even saves some duplicate work.