curl-autotools.sh: set -O3 without a local patch [ci skip]
Turns out the reason for autotools forcing its own -On option was that we
explicitly asked for it via the `--enable-optimize` ./configure option.
(The default option is `-Os`, which Daniel fixed [1][2] to be `-O2`,
upcoming in the next curl release.)

Delete `--enable-optimize` in this patch and pass `-O3` via `CFLAGS` for
one notch higher level of optimization than the default. CMake uses `-O3`
by default and we want this to be the same for all three build systems to
emit the same binaries.

With this, we are down to one local patch.

[1] https://github.com/curl/curl/commit/51083a197c883fcf28f65c932f136bc9b6e807be
[2] https://github.com/curl/curl/pull/9444