curl-autotools.sh: DLL name fixup for x86 [ci skip]
Add quotes around the `-release` option to not let the next option
spill over as a release value when `_CURL_DLL_SUFFIX_NODASH` is empty.
This happens with x86 builds, which have no DLL name suffix.

This will only make it less weird because libtool adds the dash
separator anyway. Tackle that in a separate commit.