curl-m32.sh: manually delete the .exe before build [ci skip]
In `CW_DEV_INCREMENTAL` mode there is no 'make clean' to delete it, so
the `.map` file is not re-generated, making post-build steps fail.

This bring this in-sync with `.dll`s and makes sure that the linker
commands always gets executed, to ensure the expected `.map` and `.def`
outputs are always (re-)created.