curl: move common pre-packaging code into a separate module [ci skip]
Now that `curl-m32.sh`, `curl-cmake.sh` and `curl-autotools.sh` uses
the exact same [1] post-processing / pre-packaging code, dedupe this
logic by moving it to a new file `curl-pkg.sh` and use that from the
three build-specific scripts.

[1] `curl-autotools.sh` had a special filter (to exclude `Makefile`)
when copying docs. Let's make sure to keep this logic. It's not
necessary for cmake and m32 builds, but it will not cause any harm
there.