libssh.sh: document duplicate symbol workaround with libressl [ci skip]
If `-lssh` is passed after `-lcrypto`, this happens:

```
ld.lld: error: duplicate symbol: explicit_bzero
>>> defined at libssh.a(misc.c.obj)
>>> defined at libcrypto.a(libcompatnoopt_la-explicit_bzero_win.o)
```

Reversing the order fixes it. Did not manage to build libssh without
this public symbol (it's also not an option with libressl). Forcing to
detect it in libressl works, but then the declaration is missing
throughout the build.

libssh 0.10.4, libressl 3.6.0