curl-gnumake.sh: omit OPENSSL_LIBS [ci skip]
curl's `Makefile.m32` already sets it automatically, to the same value.

Except: when dealing with pthread lib. To avoid unexpected fallout
with the shared/static override, where passing it via `LIBS` causes
the dynamic mode flag to spill into the whole lib list, including
libcurl. It's a complete accident that none of these caused a fallout
so far.

The solution would be to individually force statc/dynamic, but the `ld`
trick for it (-l:mylib) is not supported by `lld`. Nor anything else
that I know of.