curl-gnumake.sh: stop forcing static libs for z/zstd/brotli [ci skip]
This was historically necessary when building within MSYS2 to avoid
picking dynamic versions of these libs installed via MSYS2. We don't
build dynamic flavours, or we explicitly delete them to avoid linking
against them.

This change will likely restore the unwanted behaviour under MSYS2.
This will need to be fixed again and possibly fixed differently
to avoid `-Bdynamic`/`-Bstatic` on the command-line, which may have
other unwanted side-effects (unless we make sure to use it for every
`-l` option).

Also the same issue may need a fix with MSYS2 also for the autotools
and cmake builds. These have not been tested in the MSYS2 env yet.