curl-gnumake.sh: force static linking for all libs [ci skip]
Even though we don't build shared versions of dependencies (or we are
forced, we delete them), e.g. under MSYS2 the linker may still pick up
shared versions distributed as part of the environment. This happened
in the past with zlib.

To avoid this, we had some per-lib tricks to force static mode.

This patch replaces this by setting static mode for all libs. This
means passing `-Bstatic` as the first option in `LIBS`.

This in turn allows to avoid switching back and forth between the
desired and an imagined (but mostly wrong) default, with unintended
side-effects.

With such side-effects avoided, we can also drop the mitigation trick
of passing `-lpthread` via `OPENSSL_LIBS` and instead pass it via
`LIBS`.